requirements.txt
apscheduler==3.11.0
chromadb==0.5.23
fastapi==0.115.6
firebase_admin==6.6.0
python-jose
pydantic[email]
langchain_community==0.3.11
openai==1.57.3
passlib==1.7.4
praw==7.8.1
pydantic==2.10.3
pymongo==4.10.1
python-dotenv==1.0.1
pytz==2024.2
requests
tavily-python
